+++ StockScraper/BLL/Reuters_Financials_EfficienciesServices.cs
@@ -14,9 +14,9 @@ namespace BLL
         #endregion
 
         #region--Save Reuters Financials Dividends--
-        public int Save_Reuters_Financials_Efficiencies(reuters_Financials_Efficiencies rr)
+        public long Save_Reuters_Financials_Efficiencies(reuters_Financials_Efficiencies rr)
         {
-            int Reuters_FinancialsEfficiency_Id = 0;
+            long Reuters_FinancialsEfficiency_Id = 0;
             using (DBEntities db = new DBEntities())
             {
                 if (rr.Reuters_FinancialsEfficiency_Id > 0)
